cleanup macro OPENSSL_CALLBACK_FUNC
This macro provides the signature for OpenSSL callback functions.
If used consequently for all callback functions, it assures the same
signature for all them. But this is already assured by the function
declarations which use OT_OPENSSL_CALLBACK. Hence, the macro
OPENSSL_CALLBACK_FUNC will be removed which makes the function
definition more readable and allows to selectively remove unused
parameters.