Clean up usage of issueNextTransactionNum.
Problem: issueNextTransactionNum took a nym and bSaveTheNumber as
parameters. The nym is used only if bSaveTheNumber is true. This case is
a legacy case from the time where the notary assigned tx nums to nyms
without requiring them to sign for them. For the normal case, no nym is
required, but one is passed on every call anyway.
Two instances of an invocation with bSaveTheNumber=true remain, which
need a closer look. This commit cleans up the signature and invocations
of that function, and introduces a new function
`issueNextTransactionNumberToNym` to handle the two odd cases.