Renames masterCredentialID to masterID.
This in order to make it consistent in the two places the credential is
written (and read). In one place 'masterCredentialID' was used, in an
other 'masterID'.

I've opted for 'masterID' since at yet an other place 'ID' was used
signify the credential it was an attribute for (in stead of using
'credentialID'). Renaming it like so contains the change to the least
amount of places.