Dockerfile: add missing backslash
Before `cmake` was called without arguments because the backslash
was missing which is needed to indicate continuation lines in shell.

Also, CMake's `--debug-output` flag is removed because it's meant
for debugging and clutters CMake's output which leads to warning
messages being overlooked. Else, the missing backslash would have
been found sooner.