OTClient::ProcessMessageOut(): Clean up stale server connections
If an API user sends a command to one notary and then immedidately tries to send a command to a different notary, the second connection would fail because of the pre-existing connection.

Explicitly check the ID on a pre-existing connection and destroy it if it's not established with the correct notary.