Contract::DearmorAndTrim avoid spurious console output
Not every string passed to this function is an armored Contract. Frequently it's a serialized protobuf, and we don't know that it's a protobuf until this method fails.

Do not print an error message by default if decoding fails so that when protobufs are tested using this method it doesn't generate confusing console output.