Track nym derivation indices per BIP-39 seed
Previously the next BIP-32 index used for deriving a nym was tracked as a single global value.

This commit changes this behavior to instead track the next index value on a per-seed basis.

Next HD index information is no longer stored in OTWallet. If a wallet is loaded that contains that information because it was created prior to this commit, the index value is assumed to refer to the default seed and that seed's individual index is set appropriately.